Reset Button Now Working
Reset button re-loads the scene, so I was able to make scene persistent saving of the current game state and reset it when necessary.
Fixed a bug if the player quit the game mid training it would skip the training selection screen and start immediately. When in the main menu it resets the game state